backend/requirements/base.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
bcrypt>=4.0
psycopg2-binary>=2.9
python-decouple>=3.8
